Abstract UI components into `Components.js`
